Refactor peripheral manager streams
